Adds support for accepting both autosync/name and host/name SPNs on the client side Fixed an issue where the main window title did not appear until after connection Moved connection logic into ConnectionManager class Adds a UseNetTcpForLocalHost registry key to force a TCP connection rather than named pipes for connections to the local host
